Refactors pod launcher UI (#80887)
## About The Pull Request
While debugging #80857, it became immediately apparent why it was never
converted over to TypeScript: It's massive. 1203 lines in one file,
making it our largest UI! I stated in the aforementioned PR my general
guideline: 500 lines before something needs to be separated. This does
so.
## Why It's Good For The Game
- Separates a titanic UI component for easier maintainability
- Typescript conversion
## Changelog
N/A nothing player facing
