+++ tgui/packages/tgui/interfaces/CentcomPodLauncher/hooks.ts
@@ -1,0 +1,9 @@
+import { useLocalState } from '../../backend';
+
+export function useCompact() {
+  return useLocalState('compact', false);
+}
+
+export function useTab() {
+  return useLocalState('tab', 1);
+}
