+++ tgui/packages/tgui/interfaces/CentcomPodLauncher/types.ts
@@ -1,0 +1,69 @@
+import { Placement } from '@popperjs/core';
+import { BooleanLike } from 'common/react';
+
+export type PodLauncherData = {
+  bayNumber: string;
+  custom_rev_delay: number;
+  customDropoff: BooleanLike;
+  damageChoice: number;
+  defaultSoundVolume: number;
+  delays: Record<string, number>;
+  effectAnnounce: BooleanLike;
+  effectBluespace: BooleanLike;
+  effectBurst: BooleanLike;
+  effectCircle: BooleanLike;
+  effectLimb: BooleanLike;
+  effectMissile: BooleanLike;
+  effectName: number;
+  effectOrgans: BooleanLike;
+  effectQuiet: BooleanLike;
+  effectReverse: BooleanLike;
+  effectShrapnel: BooleanLike;
+  effectStealth: BooleanLike;
+  effectStun: BooleanLike;
+  effectTarget: string | null;
+  explosionChoice: number;
+  fallingSound: BooleanLike;
+  giveLauncher: BooleanLike;
+  landingSound: string | null;
+  launchChoice: number;
+  launchClone: BooleanLike;
+  launchRandomItem: BooleanLike;
+  leavingSound: string | null;
+  mapRef: string;
+  numObjects: number;
+  oldArea: string | null;
+  openingSound: string | null;
+  payload: BooleanLike;
+  picking_dropoff_turf: BooleanLike;
+  podDesc: string;
+  podName: string;
+  renderLighting: BooleanLike;
+  rev_delays: Record<string, number>;
+  reverse_option_list: Record<string, BooleanLike>;
+  shrapnelMagnitude: number;
+  shrapnelType: string;
+  soundVolume: number;
+  styleChoice: number;
+};
+
+export type PodDelay = {
+  title: string;
+  tooltip: string;
+};
+
+export type PodEffect =
+  | {
+      act: string;
+      choiceNumber?: number;
+      content?: string;
+      details?: boolean;
+      divider?: never;
+      icon: string;
+      payload?: Record<string, any>;
+      selected?: string;
+      soloSelected?: string;
+      title: string;
+      tooltipPosition?: Placement;
+    }
+  | { divider: boolean };
